Update_Logic

+++ src/main/java/org/asmjava5/service/impl/CartItemImpl.java
@@ -3,7 +3,9 @@ package org.asmjava5.service.impl;
 import lombok.RequiredArgsConstructor;
 import org.asmjava5.convert.CartItemMapstruct;
 import org.asmjava5.data.dto.request.CartItemDtoRequest;
+import org.asmjava5.data.dto.request.update.CartItemUpdateRequest;
 import org.asmjava5.data.dto.response.CartItemDtoResponse;
+import org.asmjava5.data.entity.CartItem;
 import org.asmjava5.enums.ErrorCode;
 import org.asmjava5.exception.AppException;
 import org.asmjava5.repository.CartItemRepository;
@@ -22,54 +24,36 @@ public class CartItemImpl implements CartItemService {
 
     @Override
     public List<CartItemDtoResponse> getCartItemList(String username) {
-        try {
-            var cartItemList = cartItemRepository.findByCart_User_Username(username);
-            return cartItemMapstruct.toCartItemResponseList(cartItemList);
-        } catch (RuntimeException e) {
-            throw new AppException(ErrorCode.FAIL_GET_LIST);
-        }
+        var cartItemList = cartItemRepository.findByCart_User_Username(username);
+        if (cartItemList.isEmpty()) throw new AppException(ErrorCode.FAIL_GET_LIST);
+        return cartItemMapstruct.toCartItemResponseList(cartItemList);
     }
 
     @Override
     @Transactional
     public Boolean deleteCartItemList(Long userId, List<Long> productId) {
-        try{
-            cartItemRepository.deleteByCart_User_UserIdAndProduct_ProductIdIn(userId, productId);
-            return true;
-        } catch (RuntimeException e) {
-            throw new AppException(ErrorCode.FAIL_DELETE);
-        }
+        if (!cartItemRepository.existsById(userId) && productId == null ) throw new AppException(ErrorCode.FAIL_DELETE);
+        cartItemRepository.deleteByCart_User_UserIdAndProduct_ProductIdIn(userId, productId);
+        return true;
     }
 
     @Override
     @Transactional
-    public Boolean updateCartItem(CartItemDtoRequest cartItemDtoRequest) {
-        try{
-            var cartItem = cartItemMapstruct.toCartItem(cartItemDtoRequest);
-            var check = cartItemRepository.findById(cartItem.getCartItemId());
-            if(check != null) {
-                cartItemRepository.saveAndFlush(cartItem);
-                return true;
-            }
-            return false;
-        } catch (RuntimeException e) {
-            throw new AppException(ErrorCode.FAIL_TO_SAVE_UPDATE);
-        }
+    public Boolean updateCartItem(CartItemUpdateRequest cartItemUpdateRequest) {
+        CartItem cartItem = cartItemMapstruct.toUpdateCartItem(cartItemUpdateRequest);
+        CartItem check = cartItemRepository.findById(cartItem.getCartItemId()).orElseThrow(()-> new AppException(ErrorCode.FAIL_TO_SAVE_UPDATE));
+        if (check == null) throw new AppException(ErrorCode.FAIL_TO_SAVE_UPDATE);
+        cartItemRepository.saveAndFlush(cartItem);
+        return true;
     }
 
     @Override
     @Transactional
     public Boolean addCartItem(CartItemDtoRequest cartItemDtoRequest) {
-        try{
-            var cartItem = cartItemMapstruct.toCartItem(cartItemDtoRequest);
-            var check = cartItemRepository.findById(cartItem.getCartItemId());
-            if(check == null) {
-                cartItemRepository.saveAndFlush(cartItem);
-                return true;
-            }
-            return false;
-        } catch (RuntimeException e) {
-            throw new AppException(ErrorCode.FAIL_TO_SAVE_UPDATE);
-        }
+        CartItem cartItem = cartItemMapstruct.toCartItem(cartItemDtoRequest);
+        CartItem check = cartItemRepository.findById(cartItem.getCartItemId()).orElseThrow(()-> new AppException(ErrorCode.FAIL_TO_SAVE_UPDATE));
+        if (check != null) throw new AppException(ErrorCode.FAIL_TO_SAVE_UPDATE);
+        cartItemRepository.saveAndFlush(cartItem);
+        return true;
     }
 }
